deploy/ia/agent/requirements.txt
requests>=2.28
gitpython>=3.1
